+++ test/test_list_databases_response.py
@@ -36,6 +36,7 @@ class TestListDatabasesResponse(unittest.TestCase):
         model = ListDatabasesResponse()
         if include_optional:
             return ListDatabasesResponse(
+                count = 0,
                 databases = [
                     hotdata.models.database_summary.DatabaseSummary(
                         created_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'), 
@@ -44,7 +45,10 @@ class TestListDatabasesResponse(unittest.TestCase):
                         expires_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'), 
                         id = '', 
                         name = '', )
-                    ]
+                    ],
+                has_more = True,
+                limit = 0,
+                next_cursor = ''
             )
         else:
             return ListDatabasesResponse(